// backend/src/dto/islamic-news/create-islamic-news.dto.ts
export class CreateIslamicNewsDto {
  title: string;
  description?: string;
  content?: string;
  link: string;
  image_url?: string;
  source_id?: string;
  source_name?: string;
  source_url?: string;
  language?: string;
  country?: string;
  category?: string;
  keywords?: string;
  pub_date?: Date;
  video_url?: string;
  is_archived?: boolean;
  archive_date?: Date;
}